Fixed user state bug during sign up when sign up code is used
When a user signs up with a sign up code we check if the given email matches the
sign up code email. If they match we confirm the email, but we don't reactivate
the user account when ACCOUNT_EMAIL_CONFIRMATION_REQUIRED is True.

This commit fixes this ensuring users are activated correctly when they never
receive the email confirmation by design.