Drop compatibility handling in account.compat
The `django.core.urlresolvers` module was moved to
`django.urls` in version 1.10, and deprecated in version 2.0
Since the lowest version of Django we're using here is 1.11
we can safely use functions from the latter module.

Also, as of Django 1.10, `user.is_authenticated` is a property,
and not a method:
https://docs.djangoproject.com/en/1.11/ref/contrib/auth/#django.contrib.auth.models.User.is_authenticated
So we can safely drop that part of the compatibility handling, too.